[+] se juntaron los formulario de edit y de create de usuarios ahorrando ~250 lineas de codigo

+++ src/app/features/users/user-list/user-list.ts
@@ -1,9 +1,12 @@
 import { Component, inject } from '@angular/core';
 import { HttpUsers } from '../../../core/services/http-users';
 import { AsyncPipe, JsonPipe, TitleCasePipe, UpperCasePipe } from '@angular/common';
-import { BehaviorSubject, Subscription} from 'rxjs';
+import { BehaviorSubject, Subscription } from 'rxjs';
 import { RouterLink } from '@angular/router';
 import UserCard from '../user-list-card/user-list-card';
+
+import Swal from 'sweetalert2';
+import UserForm from '../user-form/user-form';
 
 @Component({
   selector: 'app-user-list',
@@ -12,43 +15,56 @@ import UserCard from '../user-list-card/user-list-card';
   styleUrl: './user-list.css',
 })
 export default class UserList {
+  subscriberUser!: Subscription;
 
-  subscriberUser! : Subscription;
+  subscriberDeleteUser!: Subscription;
 
-  subscriberDeleteUser!: Subscription
-
-  subscriberUpdateuser!: Subscription
+  subscriberUpdateuser!: Subscription;
 
   users$ = new BehaviorSubject<any[]>([]);
 
   private httpUsers = inject(HttpUsers);
 
   ngOnInit() {
-
-    this.loadUsers()
-    
+    this.loadUsers();
   }
   onDelete(userId: string) {
-    this.subscriberDeleteUser = this.httpUsers.deleteUserbyId(userId).subscribe({
-      next: (data) => {
-        console.log(data);
-        this.loadUsers();
-      },
-      error: (error) => {
-        console.error(error);
-      },
-      complete: () => {
-        console.log('eliminado');
-      },
+    //confimacion yupiu
+
+    Swal.fire({
+      title: ' ¿ Seguro que quiere elimnar el usuario ?',
+      text: 'No se puede deshacer!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: 'Eliminado!',
+          text: 'Usuario eliminado',
+          icon: 'success',
+        });
+
+        this.subscriberDeleteUser = this.httpUsers.deleteUserbyId(userId).subscribe({
+          next: (data) => {
+            console.log(data);
+            this.loadUsers();
+          },
+          error: (error) => {
+            console.error(error);
+          },
+          complete: () => {
+            console.log('eliminado');
+          },
+        });
+      }
     });
   }
 
-  onEdit(userId: string){
+  onEdit(userId: string) {}
 
-    return console.log(userId)
-
-  }
-  
   getAllUsers() {
     return this.users$.value.length;
   }
@@ -83,20 +99,18 @@ export default class UserList {
     return number;
   }
 
-  ngOnDestroy(){
-    if (this.subscriberUser){
-
-      this.subscriberUser.unsubscribe()
-       return console.log('se elmino la sub')
-
+  ngOnDestroy() {
+    if (this.subscriberUser) {
+      this.subscriberUser.unsubscribe();
+      return console.log('se elmino la sub');
     }
-    if(this.subscriberDeleteUser){
-      this.subscriberDeleteUser.unsubscribe()
-      return console.log('se elimino la sub')
+    if (this.subscriberDeleteUser) {
+      this.subscriberDeleteUser.unsubscribe();
+      return console.log('se elimino la sub');
     }
   }
-  private loadUsers(){
-    this.subscriberUser =  this.httpUsers.getUsers().subscribe({
+  private loadUsers() {
+    this.subscriberUser = this.httpUsers.getUsers().subscribe({
       next: (data) => {
         console.log(data);
         this.users$.next(data);
@@ -108,6 +122,5 @@ export default class UserList {
         console.log('codigo funciona');
       },
     });
-
   }
 }
